Fix regex to parse callback URL & save access token
The callback URL was no longer being correctly parsed by the regex, possibly due to the fact that the callback URL from put.io might have changed at some point. Due to this, the access token is never saved and the put.io adder is stuck in a limbo waiting for authentication.